Renamed test file and removed older one
